[WIP] Debounce JenkinsCheck on the number of job failures
This is a first attempt at debouncing Jenkins checks on the
number of job failures, and not Cabot check failures.

Since the job number is stored in the status_check_result,
we can loop over the recent results, and count how many
distinct jobs have failed, and set the status to fail if
this is higher than the debounce.

However, Cabot only considers the last 10 results (hardcoded value).
Since Cabot checks the job at fairly high frequency (or at least a
frenquency higher than the Jenkins run frequency), this can mean
the status would switch to pass after 10 checks of a single check failure.